abstract common functionality and fix ros init oversight
